+++ app/(base)/new/YourDetails.tsx
@@ -6,7 +6,7 @@ import { useForm } from "react-hook-form";
 import * as z from "zod";
 import { SignedIn } from "@clerk/nextjs";
 import { PenSquare } from "lucide-react";
-import { List } from "@/server/db/schema";
+import { List } from "@/server/db/types";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import {
   FormLabel,
