+++ components/EventCard.tsx
@@ -18,7 +18,10 @@ import { ConditionalWrapper } from "./ConditionalWrapper";
 import { FollowEventDropdownButton } from "./FollowButtons";
 import { Badge } from "./ui/badge";
 import { EventWithUser } from "./EventList";
-import { eventFollows, User, Comment, EventFollow } from "@/server/db/schema";
+import { eventFollows } from "@/server/db/schema";
+import { User } from "@/server/db/types";
+import { EventFollow } from "@/server/db/types";
+import { Comment } from "@/server/db/types";
 import {
   translateToHtml,
   getDateInfoUTC,
