+++ components/EventList.tsx
@@ -1,5 +1,8 @@
 import { clsx } from "clsx";
-import { Comment, Event, EventFollow, User } from "@/server/db/schema";
+import { User } from "@/server/db/types";
+import { EventFollow } from "@/server/db/types";
+import { Event } from "@/server/db/types";
+import { Comment } from "@/server/db/types";
 import { EventCard } from "@/components/EventCard";
 import {
   Accordion,
